Restructured my code and omitted the form.js. Now each question is a form and am using useState to show the next page depending on if the user has submitted the form or not with false/true. Have also imported and used reacttooltip to help with required/validation. The default validation wasn't working on some of my form inputs so I decided to create custom ones that prompt the user with a text in a tooltip when they press the button before entering the required information.
