--- code/src/components/TypeQuestion.js
@@ -1,66 +1,0 @@
-import React, { useState, useEffect, useRef } from 'react';
-import ReactTooltip from 'react-tooltip';
-
-export const TypeQuestion = ({type, setType, errorMessage}) => {
-
-    const typesOfCraftBeer = ["Classic IPA", "Jolly Bitter", "The Belgian crafty", "Top pint Lager", "Devout Stout", "Hoppy vego (Gluten free/Vegan)"];
-
-    const [ firstRender, setFirstRender ] = useState(true);
-
-    const legendToBeFocused = useRef(null);
-
-    useEffect(() => {
-        if (firstRender) {
-            legendToBeFocused.current.focus();
-            setFirstRender(false);
-        }
-    }, [firstRender]);
-
-    /*
-    includes() checks if an array contains a certain element
-    filter() creates an array filled with all elements that pass a test implemented by the function
-    ... spread operator - Allows for array elements to be made individual and spead out in the array.
-    ? if : else
-
-    Function onTypeChange checks if type variable contains whatever is in the kinds argument which is defined in the Type.js.
-    if true -> then setType function will create an array in the type variable and not include those that are in the typeValue function
-    else -> then setType will spread out all array elements in the type variable and add whatever values that are in the typeValue into the type array. 
-    */
-    const onTypeChange = kinds => {
-        type.includes(kinds)
-        ? setType ( type.filter(item => item !== kinds) ) 
-        : setType( [...type, kinds] );
-      };
-
-    /* Mapping through the typesOfCraftBeer array to create a checkbox for each of the array elements. Saying if a checkbox is checked and includes the different elements of the array then the onChange will implement the onTypeChange function defined at the top of the page */
-    return ( 
-        <fieldset className="container">
-            <legend 
-                tabIndex="0" 
-                ref={legendToBeFocused} 
-                aria-label="Please tab and select the Craft Beer Company products you like the most and use the space bar and tab to select. Then press enter to go to the next question.">
-                    Which of the Craft Beer Co. products are your favourite?
-            </legend>
-            <div className="checkbox-row" ref={errorMessage} data-tip="Please select a type of beer">
-            <ReactTooltip backgroundColor="rgb(11, 77, 149)" class="tooltip" />
-                {typesOfCraftBeer.map((kinds) => (
-                <label 
-                    className="checkbox-container"
-                    htmlFor={kinds} 
-                    key={kinds}> 
-                    <input
-                        id={kinds}
-                        name="beerTypes"
-                        type="checkbox" 
-                        aria-label={kinds}
-                        checked={type.includes(kinds)} 
-                        onChange={() => onTypeChange(kinds)}
-                    />              
-                    {kinds}
-                    <span className="custom-checkbox"></span>                         
-                </label>
-                ))}
-            </div>
-        </fieldset> 
-    );
-};
